refactor: make Shared/Communication UI-agnostic for future Avalonia head
Keep core libs free of WPF so a later Avalonia (or other) UI can
reference them without multi-targeting Shared. WPF app behavior
should be unchanged.

Extract converter/helper logic into Shared/ConverterLogic.cs and git mv
Shared/UIUtils.cs to ECUFlasher/SharedUIUtils.cs (do not touch
ECUFlasher/UIUtils.cs baud/connection converters).

Drop UseWPF from Shared/Communication; retarget those libs plus Checksum,
ApplicationShared, and Tests to net10.0 while ECUFlasher stays
net10.0-windows.

Communication uses DescriptionAttributeConverterLogic for
programming-precondition text.

Closes #112
