+++ Shared/Shared.csproj
@@ -4,9 +4,6 @@
     <RootNamespace>ECUShared</RootNamespace>
     <AssemblyName>ECUShared</AssemblyName>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <UseWindowsForms>true</UseWindowsForms>
-    <UseWPF>true</UseWPF>
-    <ImportWindowsDesktopTargets>true</ImportWindowsDesktopTargets>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <OutputPath>bin\msil\Debug\</OutputPath>
